Move responsibility for writing in to driver
